% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCorrection.R
\name{batchCorrection}
\alias{batchCorrection}
\title{Batch-correction for CyTOF data}
\usage{
batchCorrection(sce, maxN = 1e+05, seed = 12345)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.
\code{colData(sce)} must contain file_name and batch columns.}

\item{maxN}{Numeric.
Specifies the maximum number of unique cells per batch. For example,
if one batch contains five samples, the function randomly takes \code{maxN/5}
cells from each of the samples.}

\item{seed}{Numeric. Sets a random seed.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{batchCorrection} takes an SCE object returned by \code{prepSCE} and
performs batch correction using \pkg{Harmony}.
Down-sampling is done prior to batch correction for improving speed.
}
