% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnnotation.R
\name{clusterAnnotation}
\alias{clusterAnnotation}
\title{Integration of manual cluster annotations}
\usage{
clusterAnnotation(sce, df_anno)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.
\code{colData(sce)} must contain the cluster_id column.}

\item{df_anno}{A data.frame which has the following columns: cluster_id, celltype,
celltype_detailed, and order. The order column will be used to define the order of
the factor levels for cell types.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{clusterAnnotation} takes both an SCE object returned by
\code{clusterPropagation} and a data.frame of annotations as inputs.
}
