% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\alias{clustering_with_masking}
\title{Clustering}
\usage{
clustering(
  sce,
  features = rownames(sce),
  by_exprs_values = "normexprs",
  method = "FlowSOM",
  xdim = 20,
  ydim = 20,
  maxK = 40,
  n_components = min(c(10, length(features))),
  n_neighbors = 10,
  min_dist = 0.1,
  seed = 12345
)

clustering_with_masking(
  sce,
  features = rownames(sce),
  features_masking = rownames(sce),
  by_exprs_values = "normexprs",
  method = "FlowSOM",
  xdim = 20,
  ydim = 20,
  maxK = 40,
  n_components = min(c(10, length(features))),
  n_neighbors = 10,
  min_dist = 0.1,
  seed = 12345
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{features}{A character vector.
Specifies which antigens to use for clustering.}

\item{by_exprs_values}{A character string.
Specifies which assay data to use for clustering.}

\item{method}{A character string.
Specifies which clustering method to use. Currently, either "FlowSOM" or "SNNGraph" is supported.}

\item{xdim, ydim}{Numeric. Specifies the grid size of the self-orginizing map.
For example, 20x20 grid will yield 400 clusters.}

\item{maxK}{Numeric. Specifies the maximum number of clusters to evaluate
in the metaclustering. For \code{maxK = 40}, for example,
metaclustering will be performed to obtain 40 clusters.}

\item{n_components, n_neighbors, min_dist}{Specifies parameters for UMAP and SNN-graph construction.}

\item{seed}{Numeric. Sets a random seed.}

\item{features_masking}{A character vector.
Specifies which antigens to use for the preliminary masking.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object with a
'\code{cluster_id}' column. Masked cells are assigned to the cluster 0.
}
\description{
\code{clustering} currently implements two methods. The first approach groups
cells into \code{xdim}x\code{ydim} clusters using \pkg{FlowSOM}, and then
performs metaclustering with \pkg{ConsensusClusterPlus} into \code{maxK} clusters.
The second approach performs dimention reduction through UMAP, constructs
shared nearest-neighbor graphs using \pkg{scran}, and then performs community
detection using the Leuvein algorithm in \pkg{igraph}. In both cases,
an \code{SingleCellExperiment} object with a '\code{cluster_id}' column is
returned.\cr
\code{clustering_with_masking} performs a preliminary clustering and prompts
the user to provide the clusters to be masked. The median expression
value heatmap will be generated by \code{plotClusterHeatmap}. Then the function
performs the main clustering only using the non-masked cells. The masked cells
will be assigned to a cluster ID of zero.
}
