% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPropagation.R
\name{clusterPropagation}
\alias{clusterPropagation}
\title{Batch-wise propagation of clusters through machine learning}
\usage{
clusterPropagation(
  sce,
  sce_down,
  by_exprs_values = "exprs",
  maxN = 100,
  numThreads = 4,
  seed = 12345
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.
\code{colData(sce)} must contain file_name and batch columns.}

\item{sce_down}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.
\code{colData(sce_down)} must contain columns named as file_name, batch, and
cluster_id.}

\item{by_exprs_values}{A character string.
Specifies which assay data to use for classifier training and prediction.
Must be the non-batch-corrected one in this case.}

\item{maxN}{Numeric.
Specifies the maximum number of unique cells per cluster per batch to be used
for classifier training.}

\item{numThreads}{Numeric. The number of threads for \code{extraTrees}.}

\item{seed}{Numeric. Sets a random seed.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{clusterPropagation} takes both an original SCE object returned by
\code{prepSCE} and down-sampled, batch-corrected, and clustered SCE object
returned by \code{clustering} as inputs.
This function trains batch-specific classifiers using \pkg{extraTrees},
and predicts cluster IDs in a batch-wise manner.
}
