% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampleSCE.R
\name{downsampleSCE}
\alias{downsampleSCE}
\title{Down-sampling an SCE object}
\usage{
downsampleSCE(
  sce,
  maxN = 10000,
  group_by = c("batch", "group"),
  indiv_by = NULL,
  seed = 12345
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{maxN}{Numeric.
Specifies the maximum number of unique cells per group.}

\item{group_by}{A character vector.
Specifies the colData columns for grouping.}

\item{indiv_by}{A character string.
Specifies the colData column for individual sample identifiers.}

\item{seed}{Numeric. Sets a random seed.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{downsampleSCE} down-samples an SCE object based on the grouping variables
provided. For example, when grouped by a batch variable, the function takes
maxN cells per batch. Additionally, when \code{indiv_by} is provided (e.g.,
file_name), the function takes equal numbers of cells per individual, collecting
maxN cells in total per batch.
}
