% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMUBAC.R
\docType{package}
\name{iMUBAC}
\alias{iMUBAC}
\alias{package-iMUBAC}
\title{iMUBAC: Integration of Multi-Batch Cytometry datasets.}
\description{
High-dimentional cytometry, including spectral flow cytometry and mass cytometry (CyTOF), enables deep immunophenotyping at a single-cell resolution.
  Analysis of cytometry data from multiple batches of experiments remains to be challenging due to the high-dimentionality and batch effects.
  We designed iMUBAC (Integration of Multi-Batch Cytometry datasets) to enable a rational and streamlined inter-batch comparisons through i) preprocessing, ii) batch-correction, iii) unsupervised clustering, and iv) batch-specific cell-type identification.
}
