% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterHeatmap.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Plot cluster heatmap}
\usage{
plotClusterHeatmap(
  sce,
  features = rownames(sce),
  clusters = sce$cluster_id,
  by_exprs_values = "exprs",
  fun = "median",
  scale = T,
  cluster_rows = T,
  cluster_anno = F,
  draw_dend = T,
  draw_freqs = T,
  split_by = NULL,
  hm2 = NULL
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{features}{A character vector.
Specifies which antigens to use for clustering.}

\item{clusters}{A character vector.
Specifies the cluster IDs.}

\item{by_exprs_values}{A character string.
Specifies which assay data to use for plotting.}

\item{fun}{A character string.
Specifies the function for computing the summary statistic.}

\item{scale}{Logical.
Specifies whether scaled values should be plotted.
Scaled values corresponds to cofactor arcsinh-transformed expression
values scaled between 0 and 1 using 1% and 99% percentiles as boundaries.
Note that hierarchical clustering is performed on the unscaled data.}

\item{cluster_rows}{Logical.
Specifies whether rows should be reordered by hierarchical clustering.}

\item{cluster_anno}{Logical.
Specifies whether clusters should be annotated.}

\item{draw_dend}{Logical.
Specifies if the row dendrogram should be drawn.}

\item{draw_freqs}{Logical.
Specifyies whether to display cell counts and proportions.}

\item{split_by}{A character string.
Must corresponds to a column name of \code{rowData(sce)}.
If specified, the data will be subset according to this variable,
and multiple heatmaps will be drawn.}

\item{hm2}{A character string. Specifies the right-hand side heatmap.
One of: \itemize{
\item{\code{"abundances"}: cluster frequencies across samples}
\item{a character string/vector corresponding to one/multiple marker(s):
  median marker expressions across samples and clusters}
}
Depending on argument \code{hm2}, the side heatmap can contain one of:
\itemize{
\item{relataive cluster abundances by sample}
\item{median (scaled, arcsinh-transformed) marker expressions by sample}
}}
}
\value{
a \code{\link{HeatmapList-class}} object.
}
\description{
Heatmaps summarizing the clustering result.
}
