% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDR.R
\name{plotDR}
\alias{plotDR}
\title{Plot reduced dimensions}
\usage{
plotDR(
  sce,
  dimred = "UMAP",
  colour_by = "condition",
  text_by = NULL,
  point_size = 0.5,
  alpha = 0.75,
  ...
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{dimred}{A character string specifying which dimension reduction to use.
Should be one of \code{reducedDimNames(sce)}.}

\item{colour_by}{A character string corresponding to a \code{colData(sce)} column.
Specifies the color coding. If factor, a discrete color scale will be used. Otherwise, a cividis color scale will be applied by default.}

\item{text_by}{A character string corresponding to a \code{colData(sce)} column.
Specifies the text coding.}

\item{point_size}{Numeric. Specifies the size of the dots.}

\item{alpha}{Numeric. Specifies the transparency of the dots.}

\item{...}{Additional arguments passed to \code{plotReducedDim} in \pkg{scater}.}
}
\value{
A \code{ggplot} object.
}
\description{
Dimension reduction plot colored by expression values or column metadata.
To avoid plotting too many dots as a vector graphic, \code{geom_point_rast} from \pkg{ggrastr}
is used to internally replace the \code{geom_point} layer.
}
