% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUMAP.R
\name{runUMAP}
\alias{runUMAP}
\title{Perform dimension reduction through UMAP}
\usage{
runUMAP(
  sce,
  by_exprs_values = "exprs",
  name = "UMAP",
  n_components = 2,
  n_neighbors = 100,
  min_dist = 0.5,
  scale = T,
  n_threads = parallel::detectCores(logical = F),
  seed = 12345
)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{by_exprs_values}{A character string.
Specifies which assay data to use for clustering.}

\item{name}{A character string.
Specifies the name of the reduced dimensions.}

\item{n_components, n_neighbors, min_dist, scale, n_threads}{See \code{\link[uwot]{umap}}.}

\item{seed}{Numeric. Sets a random seed.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{runUMAP} takes an SCE object and computes UMAP coordinates using \pkg{uwot}.
}
