% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMetadataSCE.R
\name{updateMetadataSCE}
\alias{updateMetadataSCE}
\title{Updating metadata of an SCE object}
\usage{
updateMetadataSCE(sce, md, merge_by = c("file_name"))
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object.}

\item{md}{A data.frame/data.table containing the following metadata:
\itemize{
\item{file_name: the file name of the underlying FCS file.}
}
  Variables that do not exist in the colData of the SCE object will be ignored.}

\item{merge_by}{A character vector.
Specifies the names of the variables used for merging.}
}
\value{
A \code{\link{SingleCellExperiment-class}} object.
}
\description{
\code{updateMetadataSCE} takes a new metadata object and updates the corresponding
columns of the colData of an SCE object.
}
